search working and ProcedureSearchBar component finished
